requirements.txt
# EdgeWord NLP Pipeline - CPU-Native Requirements

# Understanding engine (Fast-Path)
onnxruntime>=1.17.0
optimum[onnxruntime]>=1.17.0
transformers>=4.38.0

# Generation engine (Compute-Path)
# Windows pre-built wheels: pip install llama-cpp-python --prefer-binary
# AVX2 build: set CMAKE_ARGS=-DLLAMA_AVX2=on && pip install llama-cpp-python
llama-cpp-python>=0.2.57

# Utilities
numpy>=1.24.0
psutil>=5.9.0
